refactor: split code into some files
